+++ ntex/src/web/response.rs
@@ -125,9 +125,8 @@ impl WebResponse {
 
 impl From<WebResponse> for Response<Body> {
     fn from(mut res: WebResponse) -> Response<Body> {
-        let head = res.response.head_mut();
-        if res.request.head().upgrade() {
-            head.set_io(res.request.head());
+        if let Some(io) = res.request.head().take_io_rc() {
+            res.response.head_mut().set_io(io);
         }
         res.response
     }